% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrained-pipeline.R
\name{as_pipeline_model}
\alias{as_pipeline_model}
\title{Get the PipelineModel from a Spark NLP pretrained pipeline}
\usage{
as_pipeline_model(pipeline)
}
\arguments{
\item{pretrained_pipeline}{the Spark NLP PretrainedPipeline object}
}
\value{
the Spark ML pipeline model from the input
}
\description{
Spark NLP pretrained pipelines are not Spark ML pipeline models. This function
will retrieve the ML pipeline model from the pretrained pipeline object.
}
