% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{nlp_annotate_full}
\alias{nlp_annotate_full}
\title{Fully annotate some text}
\usage{
nlp_annotate_full(x, target, column = NULL)
}
\arguments{
\item{x}{some SparkNLP object that has an annotate method that takes a Spark data frame as argument}

\item{target}{the text to annotate. This can be a character string, a character vector or a data frame (with the text
in a field named "text")}

\item{column}{the column name containing text if a Spark DataFrame is passed in.}
}
\value{
If given a character vector the return value is a list of lists containing the annotations.

If given a Spark DataFrame the return value is a Spark data frame containing the annotations
}
\description{
Use Spark NLP to fully annotate some text.
}
