% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{nlp_annotation}
\alias{nlp_annotation}
\title{Spark NLP Annotation object}
\usage{
nlp_annotation(x)
}
\description{
Spark NLP Annotation object
}
\details{
A Spark NLP annotation object has the following fields:
\itemize{
\item annotatorType: the type of annotation (String)
\item begin: the index of the first character under this annotation (integer)
\item end: the index after the last character under this annotation (integer)
\item metadata: associated metadata for this annotation (Map(String, String))
\item result: the main output of the annotation (String)
\item embeddings: vector of embeddings (Array(Float))
}

See \url{https://nlp.johnsnowlabs.com/docs/en/concepts#annotation}
}
