% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tool_json_reader.R
\name{nlp_annotation_read_dataset}
\alias{nlp_annotation_read_dataset}
\title{Create a data frame from an AnnotationToolJsonReader}
\usage{
nlp_annotation_read_dataset(reader, json_path)
}
\arguments{
\item{reader}{an instance of AnnotationToolJsonReader \code{\link{nlp_annotation_tool_json_reader}}}

\item{json_path}{path to the json from annotation lab export}
}
\value{
assertion train set data frame
}
\description{
Create a data frame from an AnnotationToolJsonReader
}
