% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion_dl.R
\name{nlp_assertion_dl_pretrained}
\alias{nlp_assertion_dl_pretrained}
\title{Load a pretrained Spark NLP Assertion DL model}
\usage{
nlp_assertion_dl_pretrained(
  sc,
  input_cols,
  output_col,
  batch_size = NULL,
  max_sent_len = NULL,
  storage_ref = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{batch_size}{Parameter, which regulates the size of the batch}

\item{max_sent_len}{Parameter, which regulates the length of the longest sentence}

\item{storage_ref}{storage reference for embeddings}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{AssertionDLModel} model
}
