% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_dl.R
\name{nlp_classifier_dl}
\alias{nlp_classifier_dl}
\title{Spark NLP ClassifierDLApproach}
\usage{
nlp_classifier_dl(
  x,
  input_cols,
  output_col,
  label_col,
  batch_size = NULL,
  max_epochs = NULL,
  lr = NULL,
  dropout = NULL,
  validation_split = NULL,
  verbose = NULL,
  enable_output_logs = NULL,
  lazy_annotator = NULL,
  output_logs_path = NULL,
  uid = random_string("classifier_dl_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{label_col}{name of the column containing the category labels}

\item{batch_size}{Batch size for training}

\item{max_epochs}{Maximum number of epochs to train}

\item{lr}{Initial learning rate}

\item{dropout}{Dropout coefficient}

\item{validation_split}{proportion of data to split off for validation}

\item{verbose}{Verbosity level}

\item{enable_output_logs}{boolean to enable/disable output logs}

\item{lazy_annotator}{boolean}

\item{output_logs_path}{path to put the output logs}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML annotator that
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#classifierdl}
}
