% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextual_parser.R
\name{nlp_contextual_parser}
\alias{nlp_contextual_parser}
\title{Spark NLP ContextualParserApproach}
\usage{
nlp_contextual_parser(
  x,
  input_cols,
  output_col,
  json_path = NULL,
  dictionary = NULL,
  read_as = "TEXT",
  options = NULL,
  case_sensitive = NULL,
  prefix_and_suffix_match = NULL,
  context_match = NULL,
  update_tokenizer = NULL,
  uid = random_string("contextual_parser_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{json_path}{path to json file with rules}

\item{dictionary}{path to dictionary file in tsv or csv format}

\item{read_as}{the format of the file, can be one of {TEXT, SPARK, BINARY}.}

\item{options}{an named list containing additional parameters used when reading the dictionary file}

\item{case_sensitive}{whether to use case sensitive when matching values}

\item{prefix_and_suffix_match}{whether to force both before AND after the regex match to annotate the hit}

\item{context_match}{whether to include prior and next context to annotate the hit}

\item{update_tokenizer}{Whether to update tokenizer from pipeline when detecting multiple words on dictionary values}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that provides Regex + Contextual matching based on a JSON file
See \url{https://nlp.johnsnowlabs.com/docs/en/licensed_annotators#contextual-parser}
}
