% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date-matcher.R
\name{nlp_date_matcher}
\alias{nlp_date_matcher}
\title{Spark NLP DateMatcher}
\usage{
nlp_date_matcher(
  x,
  input_cols,
  output_col,
  format = NULL,
  uid = random_string("date_matcher_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{format}{SimpleDateFormat standard date output formatting. Defaults to yyyy/MM/dd}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that reads from different forms of date and time expressions and converts them to a provided 
date format. Extracts only ONE date per sentence. Use with sentence detector for more matches.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#datematcher}
}
