% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_normalizer.R
\name{nlp_date_normalizer}
\alias{nlp_date_normalizer}
\title{Spark NLP DateNormalizer}
\usage{
nlp_date_normalizer(
  x,
  input_cols,
  output_col,
  anchor_date_day = NULL,
  anchor_date_month = NULL,
  anchor_date_year = NULL,
  uid = random_string("date_normalizer_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{anchor_date_day}{Add an anchor day for the relative dates such as a day after tomorrow (Default: -1). By default it will use the current day. The first day of the month has value 1}

\item{anchor_date_month}{Add an anchor month for the relative dates such as a day after tomorrow (Default: -1). By default it will use the current month. Month values start from 1, so 1 stands for January.}

\item{anchor_date_year}{Add an anchor year for the relative dates such as a day after tomorrow (Default: -1). If it is not set, the by default it will use the current year. Example: 2021}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that tries to normalize dates in chunks annotations. The expected format for the
date will be YYYY/MM/DD. If the date is normalized then field normalized in metadata will be true else will
be false.
See \url{https://nlp.johnsnowlabs.com/licensed/api/com/johnsnowlabs/nlp/annotators/normalizer/DateNormalizer.html}
}
