% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency-parser.R
\name{nlp_dependency_parser}
\alias{nlp_dependency_parser}
\title{Spark NLP DependencyParserApproach}
\usage{
nlp_dependency_parser(x, input_cols, output_col, n_iterations = NULL,
  tree_bank_path = NULL, tree_bank_read_as = "TEXT",
  tree_bank_options = list(format = "text"), conll_u_path = NULL,
  conll_u_read_as = "TEXT", conll_u_options = list(format = "text"),
  uid = random_string("dependency_parser_"))
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{n_iterations}{Number of iterations in training, converges to better accuracy}

\item{tree_bank_path}{Dependency treebank folder with files in Penn Treebank format}

\item{tree_bank_read_as}{TEXT or SPARK_DATASET}

\item{tree_bank_options}{options to pass to Spark reader}

\item{conll_u_path}{Path to a file in CoNLL-U format}

\item{conll_u_read_as}{TEXT or SPARK_DATASET}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}

\item{conll_u_optiopns}{options to pass to Spark reader}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator unlabeled parser that finds a grammatical relation between two words in a sentence.
 Its input is a directory with dependency treebank files.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#dependency-parser}
}
