% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc2chunk.R
\name{nlp_doc2chunk}
\alias{nlp_doc2chunk}
\title{Spark NLP Doc2Chunk}
\usage{
nlp_doc2chunk(x, input_cols, output_col, is_array = NULL,
  chunk_col = NULL, start_col = NULL,
  start_col_by_token_index = NULL, fail_on_missing = NULL,
  lowercase = NULL, uid = random_string("doc2chunk_"))
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{is_array}{Whether the target chunkCol is ArrayType<StringType>}

\item{chunk_col}{String or StringArray column with the chunks that belong to the inputCol target}

\item{start_col}{Target INT column pointing to the token index (split by white space)}

\item{start_col_by_token_index}{Whether to use token index by whitespace or character index in startCol}

\item{fail_on_missing}{Whether to fail when a chunk is not found within inputCol}

\item{lowercase}{whether to increase matching by lowercasing everything before matching}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that Converts DOCUMENT type annotations into CHUNK type with the contents of a chunkCol. 
Chunk text must be contained within input DOCUMENT. May be either StringType or ArrayType[StringType] 
(using isArray Param) Useful for annotators that require a CHUNK type input.
See \url{https://nlp.johnsnowlabs.com/docs/en/transformers#doc2chunk}
}
