% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_logreg_classifier.R
\name{nlp_document_logreg_classifier}
\alias{nlp_document_logreg_classifier}
\title{Spark NLP DocumentLogRegClassifierApproach}
\usage{
nlp_document_logreg_classifier(
  x,
  input_cols,
  output_col,
  fit_intercept = NULL,
  label_column = NULL,
  labels = NULL,
  max_iter = NULL,
  merge_chunks = NULL,
  tol = NULL,
  uid = random_string("document_logreg_classifier_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{fit_intercept}{whether to fit an intercept term (Default: true)}

\item{label_column}{column with the value result we are trying to predict.}

\item{labels}{array to output the label in the original form.}

\item{max_iter}{maximum number of iterations (Default: 10)}

\item{merge_chunks}{whether to merge all chunks in a document or not (Default: false)}

\item{tol}{convergence tolerance after each iteration (Default: 1e-6)}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that
See \url{https://nlp.johnsnowlabs.com/licensed/api/com/johnsnowlabs/nlp/annotators/classification/DocumentLogRegClassifierApproach.html}
}
