% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_normalizer.R
\name{nlp_document_normalizer}
\alias{nlp_document_normalizer}
\title{Spark NLP DocumentNormalizer}
\usage{
nlp_document_normalizer(
  x,
  input_cols,
  output_col,
  action = NULL,
  encoding = NULL,
  lower_case = NULL,
  patterns = NULL,
  policy = NULL,
  replacement = NULL,
  uid = random_string("document_normalizer_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{action}{Action to perform applying regex patterns on text}

\item{encoding}{File encoding to apply on normalized documents (Default: "disable")}

\item{lower_case}{Whether to convert strings to lowercase (Default: false)}

\item{patterns}{Normalization regex patterns which match will be removed from document (Default: Array("<\link{^>}*>"))}

\item{policy}{RemovalPolicy to remove patterns from text with a given policy (Default: "pretty_all"). Possible values are "all", "pretty_all", "first", "pretty_first"}

\item{replacement}{Replacement string to apply when regexes match (Default: " ")}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer which normalizes raw text from tagged text, e.g.
scraped web pages or xml documents, from document type columns into Sentence.
Removes all dirty characters from text following one or more input regex patterns.
Can apply not wanted character removal with a specific policy. Can apply lower case normalization.
}
\details{
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#documentnormalizer-text-cleaning}
}
