% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elmo-embeddings.R
\name{nlp_elmo_embeddings_pretrained}
\alias{nlp_elmo_embeddings_pretrained}
\title{Load a pretrained Spark NLP ElmoEmbeddings model}
\usage{
nlp_elmo_embeddings_pretrained(
  sc,
  input_cols,
  output_col,
  case_sensitive = NULL,
  batch_size = NULL,
  dimension = NULL,
  pooling_layer = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{case_sensitive}{whether to treat the tokens as case insensitive when looking up their embedding}

\item{batch_size}{batch size}

\item{dimension}{the embedding dimension}

\item{pooling_layer}{word_emb, lstm_outputs1, lstm_outputs2 or elmo}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{ElmoEmbeddings} model
}
