% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings_finisher.R
\name{nlp_embeddings_finisher}
\alias{nlp_embeddings_finisher}
\title{Spark NLP EmbeddingsFinisher}
\usage{
nlp_embeddings_finisher(
  x,
  input_cols,
  output_cols,
  clean_annotations = NULL,
  output_as_vector = NULL,
  uid = random_string("embeddings_finisher_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_cols}{Output columns. String array.}

\item{clean_annotations}{Whether to remove and cleanup the rest of the annotators (columns)}

\item{output_as_vector}{if enabled, it will output the embeddings as Vectors instead of arrays}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that is designed to deal with embedding annotators: WordEmbeddings, BertEmbeddings, 
SentenceEmbeddingd, and ChunkEmbeddings. By using EmbeddingsFinisher you can easily transform your embeddings 
into array of floats or Vectors which are compatible with Spark ML functions such as LDA, K-mean, Random Forest
 classifier or any other functions that require featureCol.
See \url{https://nlp.johnsnowlabs.com/docs/en/transformers#embeddingsfinisher}
}
