% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entity_ruler.R
\name{nlp_entity_ruler}
\alias{nlp_entity_ruler}
\title{Spark NLP EntityRulerApproach}
\usage{
nlp_entity_ruler(
  x,
  input_cols,
  output_col,
  case_sensitive = NULL,
  enable_pattern_regex = NULL,
  patterns_resource_path = NULL,
  patterns_resource_read_as = NULL,
  patterns_resource_options = NULL,
  storage_path = NULL,
  storage_ref = NULL,
  use_storage = NULL,
  uid = random_string("entity_ruler_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{case_sensitive}{Whether to ignore case in index lookups (Default depends on model)}

\item{enable_pattern_regex}{Enables regex pattern match (Default: false).}

\item{patterns_resource_path}{Resource in JSON or CSV format to map entities to patterns (Default: null).}

\item{patterns_resource_read_as}{TEXT or SPARK_DATASET}

\item{patterns_resource_options}{options passed to the reader. (Default: list("format" = "JSON"))}

\item{storage_path}{Path to the external resource.}

\item{storage_ref}{Unique identifier for storage (Default: this.uid)}

\item{use_storage}{Whether to use RocksDB storage to serialize patterns (Default: true).}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#entityruler}
}
