% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tool_json_reader.R
\name{nlp_generate_assertion_train_set}
\alias{nlp_generate_assertion_train_set}
\title{Generate an assertion training set from an AnnotationToolJsonReader}
\usage{
nlp_generate_assertion_train_set(
  reader,
  df,
  sentence_col = "sentence",
  assertion_col = "assertion_label"
)
}
\arguments{
\item{reader}{an instance of AnnotationToolJsonReader \code{\link{nlp_annotation_tool_json_reader}}}

\item{df}{a Spark Dataframe}

\item{sentence_col}{the name of the sentence column}

\item{assertion_col}{the name of the assertion column}
}
\value{
assertion training set data frame
}
\description{
Generate an assertion training set from an AnnotationToolJsonReader
}
