% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tool_json_reader.R
\name{nlp_generate_colln}
\alias{nlp_generate_colln}
\title{Generate a CoNLL format file from a data frame using an AnnotationToolJsonReader}
\usage{
nlp_generate_colln(
  reader,
  df,
  path,
  task_col = "task_id",
  token_col = "token",
  ner_label = "ner_label"
)
}
\arguments{
\item{reader}{an instance of AnnotationToolJsonReader \code{\link{nlp_annotation_tool_json_reader}}}

\item{df}{a Spark Dataframe}

\item{task_col}{the name of the task column}

\item{token_col}{the name of the token column}

\item{ner_label}{the name of the ner label column}
}
\description{
Generate a CoNLL format file from a data frame using an AnnotationToolJsonReader
}
