% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_tool_json_reader.R
\name{nlp_generate_plain_assertion_train_set}
\alias{nlp_generate_plain_assertion_train_set}
\title{Generate a plain assertion training set from an AnnotationToolJsonReader}
\usage{
nlp_generate_plain_assertion_train_set(
  reader,
  df,
  task_col = "task_id",
  token_col = "token",
  ner_label = "ner_label",
  assertion_label = "assertion_label"
)
}
\arguments{
\item{reader}{an instance of AnnotationToolJsonReader \code{\link{nlp_annotation_tool_json_reader}}}

\item{df}{a Spark Dataframe}

\item{task_col}{the name of the task column}

\item{token_col}{the name of the token column}

\item{ner_label}{the name of the ner label column}

\item{assertion_col}{the name of the assertion column}
}
\value{
assertion training set data frame
}
\description{
Generate a plain assertion training set from an AnnotationToolJsonReader
}
