% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlp_get_classes}
\alias{nlp_get_classes}
\title{Get classes used to train a model}
\usage{
nlp_get_classes(model)
}
\arguments{
\item{model}{a trained SparkNLP model that implements getClasses()}
}
\value{
a list of classes
}
\description{
Get classes used to train a model
}
