% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph-finisher.R
\name{nlp_graph_finisher}
\alias{nlp_graph_finisher}
\title{Spark NLP GraphFinisher}
\usage{
nlp_graph_finisher(
  x,
  input_col,
  output_col,
  clean_annotations = NULL,
  include_metadata = NULL,
  output_as_array = NULL,
  uid = random_string("graph_finisher_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_col}{Input column. String.}

\item{output_col}{Output column. String.}

\item{clean_annotations}{Whether to remove annotation columns (Default: true)}

\item{include_metadata}{Annotation metadata format (Default: false)}

\item{output_as_array}{Finisher generates an Array with the results instead of string (Default: true)}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Helper class to convert the knowledge graph from GraphExtraction into a generic format, such as RDF.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#graphfinisher}
}
