% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language-detector-dl.R
\name{nlp_language_detector_dl_pretrained}
\alias{nlp_language_detector_dl_pretrained}
\title{Load a pretrained Spark NLP LanguageDetectorDL model}
\usage{
nlp_language_detector_dl_pretrained(
  sc,
  input_cols,
  output_col,
  alphabet = NULL,
  coalesce_sentences = NULL,
  language = NULL,
  threshold = NULL,
  threshold_label = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{alphabet}{alphabet used to feed the TensorFlow model for prediction (Map of string to integer) This should be an R environment}

\item{coalesce_sentences}{If sets to true the output of all sentences will be averaged to one output instead of one output per sentence. (boolean)}

\item{language}{used to map prediction to two-letter (ISO 639-1) language codes (Map of string to integer) This should be an R environment}

\item{threshold}{The minimum threshold for the final result otheriwse it will be either Unknown or the value set in thresholdLabel.}

\item{threshold_label}{In case the score of prediction is less than threshold, what should be the label.}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{LanguageDetectorDL} model
}
