% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemmatizer.R
\name{nlp_lemmatizer}
\alias{nlp_lemmatizer}
\title{Spark NLP Lemmatizer}
\usage{
nlp_lemmatizer(x, input_cols, output_col, dictionary_path = NULL,
  dictionary_key_delimiter = "->", dictionary_value_delimiter = "\\t",
  dictionary_read_as = "LINE_BY_LINE", dictionary_options = list(format
  = "text"), uid = random_string("lemmatizer_"))
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{dictionary_path}{Path to lemma dictionary, in lemma vs possible words format.}

\item{dictionary_key_delimiter}{key delimiter in the dictionary file}

\item{dictionary_value_delimiter}{value delimiter in the dictionary file}

\item{dictionary_read_as}{readAs LINE_BY_LINE or SPARK_DATASET}

\item{dictionary_options}{options passed to the spark reader if read_as is SPARK_DATASET}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  a default pretrained NLP model appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that retrieves lemmas out of words with the objective of returning a base dictionary word
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#lemmatizer}
}
