% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings-helper.R
\name{nlp_load_embeddings}
\alias{nlp_load_embeddings}
\title{Spark NLP EmbeddingsHelper load}
\usage{
nlp_load_embeddings(sc, path, format, reference, dims,
  case_sensitive = FALSE)
}
\arguments{
\item{sc}{Spark connection}

\item{path}{path to the embeddings file(s)}

\item{format}{TEXT, BINARY, SPARKNLP}

\item{reference}{the name to give the embeddings so they can be referred to later}

\item{dims}{dimensions of the vectors}

\item{case_sensitive}{case sensitivity of the embeddings}
}
\value{
a handle to the ClusterWordEmbeddings object in Spark
}
\description{
Uses EmbedddingsHelper to load word embeddings for annotators. 
See \url{https://nlp.johnsnowlabs.com/docs/en/concepts#embeddingshelper}
}
