% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marian_transformer.R
\name{nlp_marian_transformer}
\alias{nlp_marian_transformer}
\title{Spark NLP MarianTransformer}
\usage{
nlp_marian_transformer(
  x,
  input_cols,
  output_col,
  lang_id = NULL,
  max_input_length = NULL,
  max_output_length = NULL,
  vocabulary = NULL,
  uid = random_string("marian_transformer_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{lang_id}{A string representing the target language in the form of >>id<< (id = valid target language ID)}

\item{max_input_length}{Controls the maximum length for encoder inputs (source language texts) Default: 40}

\item{max_output_length}{Controls the maximum length for decoder outputs (target language texts) Default: 40}

\item{vocabulary}{Vocabulary used to encode and decode piece tokens generated by SentencePiece This will be set once the model is created and cannot be changed afterwards}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that
See \url{https://nlp.johnsnowlabs.com/api/#com.johnsnowlabs.nlp.annotators.seq2seq.MarianTransformer}
}
