% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_classifier_dl.R
\name{nlp_multi_classifier_dl}
\alias{nlp_multi_classifier_dl}
\title{Spark NLP MultiClassifierDLApproach}
\usage{
nlp_multi_classifier_dl(
  x,
  input_cols,
  output_col,
  batch_size = NULL,
  enable_output_logs = NULL,
  label_col = NULL,
  lr = NULL,
  max_epochs = NULL,
  output_logs_path = NULL,
  shuffle_per_epoch = NULL,
  threshold = NULL,
  validation_split = NULL,
  verbose = NULL,
  uid = random_string("multi_classifier_dl_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{batch_size}{Batch size}

\item{enable_output_logs}{whether to output to annotators log folder}

\item{label_col}{column with label per each document}

\item{lr}{learning rate}

\item{max_epochs}{maximum number of epoch to train}

\item{output_logs_path}{output logs path}

\item{shuffle_per_epoch}{shuffle per epoch}

\item{threshold}{the minimum threshold for each label to be accepted}

\item{validation_split}{choose the proportion of training dataset to be validated against the model on each epoch}

\item{verbose}{level of verbosity during training (integer)}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#multiclassifierdl-multi-label-text-classification}
}
