% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_date-matcher.R
\name{nlp_multi_date_matcher}
\alias{nlp_multi_date_matcher}
\title{Spark NLP MultiDateMatcher}
\usage{
nlp_multi_date_matcher(
  x,
  input_cols,
  output_col,
  anchor_date_day = NULL,
  anchor_date_month = NULL,
  anchor_date_year = NULL,
  default_day_when_missing = NULL,
  read_month_first = NULL,
  format = NULL,
  source_language = NULL,
  uid = random_string("multi_date_matcher_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{anchor_date_day}{Add an anchor day for the relative dates such as a day after tomorrow (Default: -1).
By default it will use the current day. The first day of the month has value 1.}

\item{anchor_date_month}{Add an anchor month for the relative dates such as a day after tomorrow (Default: -1).
By default it will use the current month. Month values start from 1, so 1 stands for January.}

\item{anchor_date_year}{Add an anchor year for the relative dates such as a day after tomorrow (Default: -1).
If it is not set, the by default it will use the current year. Example: 2021}

\item{default_day_when_missing}{Which day to set when it is missing from parsed input (Default: 1)}

\item{read_month_first}{Whether to interpret dates as MM/DD/YYYY instead of DD/MM/YYYY (Default: true)}

\item{format}{Output format of parsed date. Defaults to yyyy/MM/dd}

\item{source_language}{Source language for explicit translation}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that reads from different forms of date and time expressions and converts them to a provided
date format. E
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#multidatematcher}
}
