% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ner-dl.R
\name{nlp_ner_dl}
\alias{nlp_ner_dl}
\title{Spark NLP NerDLApproach Named Entity Recognition Deep Learning annotator}
\usage{
nlp_ner_dl(
  x,
  input_cols,
  output_col,
  label_col = NULL,
  max_epochs = NULL,
  lr = NULL,
  po = NULL,
  batch_size = NULL,
  dropout = NULL,
  verbose = NULL,
  include_confidence = NULL,
  random_seed = NULL,
  uid = random_string("ner_dl_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{label_col}{If DatasetPath is not provided, this Seq[Annotation] type of column should have labeled data per token}

\item{max_epochs}{Maximum number of epochs to train}

\item{lr}{Initial learning rate}

\item{po}{Learning rate decay coefficient. Real Learning Rate: lr / (1 + po * epoch)}

\item{batch_size}{Batch size for training}

\item{dropout}{Dropout coefficient}

\item{verbose}{Verbosity level}

\item{include_confidence}{whether to include confidence values}

\item{random_seed}{Random seed}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}

When \code{x} is a \code{spark_connection} the function returns a NerDLApproach estimator.
When \code{x} is a \code{ml_pipeline} the pipeline with the NerDLApproach added. When \code{x}
is a \code{tbl_spark} a transformed \code{tbl_spark}  (note that the Dataframe passed in must have the input_cols specified).
}
\description{
This Named Entity recognition annotator allows to train generic NER model based on Neural Networks. 
Its train data (train_ner) is either a labeled or an external CoNLL 2003 IOB based spark dataset with Annotations 
columns. Also the user has to provide word embeddings annotation column.
}
\details{
Neural Network architecture is Char CNNs - BiLSTM - CRF that achieves state-of-the-art in most datasets.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#ner-dl}
}
