% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{nlp_normalizer}
\alias{nlp_normalizer}
\title{Spark NLP Normalizer}
\usage{
nlp_normalizer(
  x,
  input_cols,
  output_col,
  cleanup_patterns = NULL,
  lowercase = NULL,
  dictionary_path = NULL,
  dictionary_delimiter = NULL,
  dictionary_read_as = "LINE_BY_LINE",
  dictionary_options = list(format = "text"),
  uid = random_string("normalizer_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{cleanup_patterns}{Regular expressions list for normalization, defaults [^A-Za-z]}

\item{lowercase}{lowercase tokens, default true}

\item{dictionary_path}{txt file with delimited words to be transformed into something else}

\item{dictionary_delimiter}{delimiter of the dictionary text file}

\item{dictionary_read_as}{LINE_BY_LINE or SPARK_DATASET}

\item{dictionary_options}{options to pass to the Spark reader}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  a default pretrained NLP model appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that removes all dirty characters from text following a regex pattern and transforms words based on a provided dictionary
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#normalizer}
}
