% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron-model.R
\name{nlp_perceptron_model}
\alias{nlp_perceptron_model}
\title{Spark NLP PerceptronModel}
\usage{
nlp_perceptron_model(x, input_cols, output_col, n_iterations = NULL,
  pos_column = NULL, uid = random_string("perceptron_model_"))
}
\arguments{
\item{x}{}

\item{input_cols}{}

\item{output_col}{}

\item{n_iterations}{}

\item{pos_column}{}

\item{uid}{}
}
\value{
When \code{x} is a \code{spark_connection} the function returns a PerceptronModel transformer.
When \code{x} is a \code{ml_pipeline} the pipeline with the PerceptronModel added. When \code{x}
is a \code{tbl_spark} a transformed \code{tbl_spark}  (note that the Dataframe passed in must have the input_cols specified).
}
\description{
Spark ML transformer that 
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#postagger}
}
