% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perceptron.R
\name{nlp_perceptron_pretrained}
\alias{nlp_perceptron_pretrained}
\title{Load a pretrained Spark NLP model}
\usage{
nlp_perceptron_pretrained(sc, name = NULL, lang = NULL,
  remote_loc = NULL)
}
\arguments{
\item{sc}{A Spark connection}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{Perceptron} model
}
