% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretrained-pipeline.R
\name{nlp_pretrained_pipeline}
\alias{nlp_pretrained_pipeline}
\title{Spark NLP Pretrained pipeline}
\usage{
nlp_pretrained_pipeline(x, download_name, lang = "en",
  source = "public/models", parse_embeddings_vectors = FALSE)
}
\arguments{
\item{x}{a Spark connection, Spark dataframe or string or character vector}

\item{download_name}{the name of the pretrained pipeline to download and create}

\item{lang}{the language of the pipeline}

\item{source}{the source for the pipeline file}

\item{parse_embeddings_vectors}{whether to parse the embeddings vectors or not}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of 
  a \code{ml_pipeline} created from the pretrained pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, a the pretrained pipeline is created and immediately 
  run on the provied dataframe using \code{ml_fit_and_transform} returning the transformed data frame.
}
}
\description{
Creates a Spark NLP pretrained pipeline. See 
\url{https://nlp.johnsnowlabs.com/api/index.html#com.johnsnowlabs.nlp.pretrained.PretrainedPipeline} for the 
default values for the parameters if left null
}
