% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlp_pubtator_read_dataset}
\alias{nlp_pubtator_read_dataset}
\title{PubTator Dataset}
\usage{
nlp_pubtator_read_dataset(sc, path)
}
\arguments{
\item{sc}{Spark connection}

\item{path}{path to a PubTator file}
}
\value{
Spark Dataframe created from the PubTator file
}
\description{
The PubTator format includes medical papers’ titles, abstracts, and tagged chunks
(see \href{http://bioportal.bioontology.org/ontologies/EDAM?p=classes&conceptid=format_3783}{PubTator Docs} and
\href{http://github.com/chanzuckerberg/MedMentions}{MedMentions Docs}
for more information). We can create a Spark DataFrame from a PubTator text file.
}
