% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_ner_chunks_filter.R
\name{nlp_re_ner_chunks_filter}
\alias{nlp_re_ner_chunks_filter}
\title{Spark NLP RENerChunksFilter}
\usage{
nlp_re_ner_chunks_filter(
  x,
  input_cols,
  output_col,
  max_syntactic_distance = NULL,
  relation_pairs = NULL,
  uid = random_string("re_ner_chunks_filter_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{max_syntactic_distance}{Maximal syntactic distance, as threshold (Default: 0)}

\item{relation_pairs}{List of dash-separated pairs of named entities
("ENTITY1-ENTITY2", e.g. "Biomarker-RelativeDay"), which will be processed}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that filters and outputs combinations of relations between
extracted entities, for further processing. This annotator is especially useful
to create inputs for the RelationExtractionDLModel.
}
\details{
See \url{https://nlp.johnsnowlabs.com/docs/en/licensed_annotators#renerchunksfilter}
}
