% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursive-tokenizer.R
\name{nlp_recursive_tokenizer}
\alias{nlp_recursive_tokenizer}
\title{Spark NLP RecursiveTokenizer}
\usage{
nlp_recursive_tokenizer(
  x,
  input_cols,
  output_col,
  infixes = NULL,
  prefixes = NULL,
  suffixes = NULL,
  white_list = NULL,
  uid = random_string("recursive_tokenizer_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{infixes}{strings that will be split when found at the middle of a token}

\item{prefixes}{strings that will be split when found at the beginning of a token}

\item{suffixes}{strings that will be split when found at the end of a token}

\item{white_list}{whitelist}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML model that tokenizes
See \url{https://nlp.johnsnowlabs.com/api/index#com.johnsnowlabs.nlp.annotators.RecursiveTokenizer}
}
