% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_matcher.R
\name{nlp_regex_matcher}
\alias{nlp_regex_matcher}
\title{Spark NLP RegexMatcher}
\usage{
nlp_regex_matcher(
  x,
  input_cols,
  output_col,
  strategy = NULL,
  rules_path = NULL,
  rules_path_delimiter = ",",
  rules_path_read_as = "TEXT",
  rules_path_options = list(format = "text"),
  uid = random_string("regex_matcher_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{strategy}{Can be any of MATCH_FIRST|MATCH_ALL|MATCH_COMPLETE}

\item{rules_path}{Path to file containing a set of regex,key pair}

\item{rules_path_delimiter}{delimiter between regex and key in the file}

\item{rules_path_read_as}{TEXT or SPARK_DATASET}

\item{rules_path_options}{options passed to Spark reader if read_as is SPARK_DATASET}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that 
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#regexmatcher}
}
