% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation_extraction.R
\name{nlp_relation_extraction}
\alias{nlp_relation_extraction}
\title{Spark NLP RelationExtractionApproach}
\usage{
nlp_relation_extraction(
  x,
  input_cols,
  output_col,
  batch_size = NULL,
  dropout = NULL,
  epochs_number = NULL,
  feature_scaling = NULL,
  fix_imbalance = NULL,
  from_entity_begin_col = NULL,
  from_entity_end_col = NULL,
  from_entity_label_col = NULL,
  label_col = NULL,
  learning_rate = NULL,
  model_file = NULL,
  output_logs_path = NULL,
  to_entity_begin_col = NULL,
  to_entity_end_col = NULL,
  to_entity_label_col = NULL,
  validation_split = NULL,
  uid = random_string("relation_extraction_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{batch_size}{batch size}

\item{dropout}{dropout coefficient}

\item{epochs_number}{maximum number of epochs to train}

\item{feature_scaling}{feature scaling method}

\item{fix_imbalance}{Fix the imbalance in the training set by replicating examples of under represented categories}

\item{from_entity_begin_col}{Column for beginning of 'from' entity}

\item{from_entity_end_col}{Column for end of 'from' entity}

\item{from_entity_label_col}{Column for 'from' entity label}

\item{label_col}{Column with label per each document}

\item{learning_rate}{learning rate}

\item{model_file}{location of file of the model used for classification}

\item{output_logs_path}{path to folder to output logs}

\item{to_entity_begin_col}{Column for beginning of 'to' entity}

\item{to_entity_end_col}{Column for end of 'to' entity}

\item{to_entity_label_col}{Column for 'to' entity label}

\item{validation_split}{Choose the proportion of training dataset to be validated against the model on each Epoch.}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that trains a TensorFlow model for relation extraction.
The Tensorflow graph in .pb format needs to be specified with setModelFile.
The result is a RelationExtractionModel. To start training, see the parameters
that need to be set in the Parameters section.

See \url{https://nlp.johnsnowlabs.com/docs/en/licensed_annotators#relationextraction}
}
