% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relation_extraction.R
\name{nlp_relation_extraction_pretrained}
\alias{nlp_relation_extraction_pretrained}
\title{Load a pretrained Spark NLP Relation Extraction model}
\usage{
nlp_relation_extraction_pretrained(
  sc,
  input_cols,
  output_col,
  relation_pairs,
  max_syntactic_distance = NULL,
  feature_scaling = NULL,
  prediction_threshold = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{relation_pairs}{List of dash-separated pairs of named entities ("ENTITY1-ENTITY2",
e.g. "Biomarker-RelativeDay"), which will be processed}

\item{max_syntactic_distance}{Maximal syntactic distance, as threshold (Default: 0)}

\item{feature_scaling}{Feature scaling method.}

\item{prediction_threshold}{Minimal activation of the target unit to encode a new relation instance (Default: 0.5f)}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{RelationExtractionModel} model
}
