% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resource_downloader.R
\name{nlp_resource_downloader}
\alias{nlp_resource_downloader}
\alias{ResourceDownloader}
\alias{nlp_show_public_pipelines}
\alias{nlp_show_public_models}
\alias{nlp_clear_cache}
\alias{nlp_show_available_annotators}
\title{SparkNLP ResourceDownloader functions}
\usage{
nlp_show_public_pipelines(sc, lang = NULL, version = NULL)

nlp_show_public_models(sc, annotator = NULL, lang = NULL, version = NULL)

nlp_clear_cache(sc, name = NULL, language = NULL, remote_loc = NULL)

nlp_show_available_annotators(sc)
}
\arguments{
\item{sc}{a spark_connect object}

\item{lang}{language to restrict the results to}

\item{version}{Spark NLP version to restrict results to}

\item{annotator}{name of annotator to restrict results}

\item{name}{name of object to clear}

\item{language}{language to clear}

\item{remote_loc}{remote_loc of models to clear}
}
\value{
a markdown table containing the models or pipelines filtered by the provided arguments
}
\description{
ResourceDownloader provides functions to easily look for pretrained models & pipelines
inside Spark NLP. You can filter models or pipelines via language, version,
or the name of the annotator
}
