% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/embeddings-helper.R
\name{nlp_save_embeddings}
\alias{nlp_save_embeddings}
\title{Spark NLP EmbeddingsHelper save}
\usage{
nlp_save_embeddings(sc, path, embeddings)
}
\arguments{
\item{sc}{Spark connection}

\item{path}{path to the embeddings file(s)}

\item{embeddings}{the embeddings to save. Should be a ClusterWordEmbeddings object}
}
\value{
a handle to the ClusterWordEmbeddings object in Spark
}
\description{
Uses EmbedddingsHelper to load word embeddings for annotators. 
See \url{https://nlp.johnsnowlabs.com/docs/en/concepts#embeddingshelper}
}
