% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence-detector.R
\name{nlp_sentence_detector}
\alias{nlp_sentence_detector}
\title{Spark NLP SentenceDetector - sentence boundary detector}
\usage{
nlp_sentence_detector(
  x,
  input_cols,
  output_col,
  custom_bounds = NULL,
  use_custom_only = NULL,
  use_abbreviations = NULL,
  explode_sentences = NULL,
  detect_lists = NULL,
  min_length = NULL,
  max_length = NULL,
  split_length = NULL,
  uid = random_string("sentence_detector_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{custom_bounds}{Custom sentence separator text. Optional.}

\item{use_custom_only}{Use only custom bounds without considering those of Pragmatic Segmenter. Defaults to false. Needs customBounds.}

\item{use_abbreviations}{Whether to consider abbreviation strategies for better accuracy but slower performance. Defaults to true.}

\item{explode_sentences}{Whether to split sentences into different Dataset rows. Useful for higher parallelism in fat rows. Defaults to false.}

\item{detect_lists}{whether to take lists into consideration at sentence detection}

\item{min_length}{set the minimum allowed length for each sentence}

\item{max_length}{set the maximum allowed length for each sentence}

\item{split_length}{length at which sentences will be forcibly set}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML Transformer that finds sentence bounds in raw text. Applies rule from Pragmatic Segmenter
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#sentencedetector}
}
