% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_detector_dl.R
\name{nlp_sentence_detector_dl}
\alias{nlp_sentence_detector_dl}
\title{Spark NLP SentenceDetectorDLApproach}
\usage{
nlp_sentence_detector_dl(
  x,
  input_cols,
  output_col,
  epochs_number = NULL,
  impossible_penultimates = NULL,
  model = NULL,
  output_logs_path = NULL,
  validation_split = NULL,
  uid = random_string("sentence_detector_dl_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{epochs_number}{maximum number of epochs to train}

\item{impossible_penultimates}{impossible penultimates}

\item{model}{model architecture}

\item{output_logs_path}{path to folder to output logs}

\item{validation_split}{choose the proportion of training dataset to be validated agaisnt the model on each epoch}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators}
}
