% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentence_entity_resolver.R
\name{nlp_sentence_entity_resolver_pretrained}
\alias{nlp_sentence_entity_resolver_pretrained}
\title{Load a pretrained Spark NLP T5 Transformer model}
\usage{
nlp_sentence_entity_resolver_pretrained(
  sc,
  input_cols,
  output_col,
  case_sensitive = NULL,
  confidence_function = NULL,
  distance_function = NULL,
  extra_mass_penalty = NULL,
  miss_as_empty = NULL,
  neighbors = NULL,
  threshold = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{case_sensitive}{whether to treat the entities as case sensitive}

\item{confidence_function}{what function to use to calculate confidence: INVERSE or SOFTMAX}

\item{distance_function}{what distance function to use for KNN: 'EUCLIDEAN' or 'COSINE'}

\item{miss_as_empty}{whether or not to return an empty annotation on unmatched chunks}

\item{neighbors}{number of neighbours to consider in the KNN query to calculate WMD}

\item{threshold}{threshold value for the aggregated distance#'}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{T5TransformerModel} model
}
