% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment-detector.R
\name{nlp_sentiment_detector}
\alias{nlp_sentiment_detector}
\title{Spark NLP SentimentDetector}
\usage{
nlp_sentiment_detector(x, input_cols, output_col,
  decrement_multiplier = NULL, dictionary_path,
  dictionary_delimiter = ",", dictionary_read_as = "LINE_BY_LINE",
  dictionary_options = list(format = "text"), enable_score = NULL,
  increment_multiplier = NULL, negative_multiplier = NULL,
  positive_multiplier = NULL, reverse_multiplier = NULL,
  uid = random_string("sentiment_detector_"))
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{decrement_multiplier}{defaults to -2.0}

\item{dictionary_path}{path to file with list of inputs and their content}

\item{dictionary_delimiter}{delimiter in dictionary file}

\item{dictionary_read_as}{LINE_BY_LINE or SPARK_DATASET}

\item{dictionary_options}{options to pass to the Spark reader. Defaults to {"format" = "text"}}

\item{enable_score}{}

\item{increment_multiplier}{defaults to 2.0}

\item{negative_multiplier}{defaults to -1.0}

\item{positive_multiplier}{defaults to 1.0}

\item{reverse_multiplier}{defaults to -1.0}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that scores a sentence for a sentiment
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#sentimentdetector}
}
