% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlp_set_input_cols}
\alias{nlp_set_input_cols}
\title{Set the input column names}
\usage{
nlp_set_input_cols(jobj, input_cols)
}
\arguments{
\item{jobj}{the object setting the input columns on}

\item{input_cols}{the input column names}
}
\description{
Set the input column names
}
