% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlp_set_output_col}
\alias{nlp_set_output_col}
\title{Set the output column name}
\usage{
nlp_set_output_col(jobj, output_col)
}
\arguments{
\item{jobj}{the object setting the input columns on}

\item{output_col}{the input column names}
}
\description{
Set the output column name
}
