% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlp_set_param}
\alias{nlp_set_param}
\title{Set a parameter on an NLP model object}
\usage{
nlp_set_param(x, param, value)
}
\arguments{
\item{x}{A Spark NLP object, either a pipeline stage or an annotator}

\item{param}{The parameter to set}

\item{value}{The value to use when setting the parameter}
}
\value{
the NLP model object with the parameter set
}
\description{
Set a parameter on an NLP model object
}
