% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nlp_set_param_tuple2}
\alias{nlp_set_param_tuple2}
\title{Set a Tuple2 parameter on an NLP model object}
\usage{
nlp_set_param_tuple2(x, param, value)
}
\arguments{
\item{x}{A Spark NLP object, either a pipeline stage or an annotator}

\item{param}{The parameter to set}

\item{value}{The value to use when setting the parameter. This should be a list of size 2}
}
\value{
the NLP model object with the parameter set
}
\description{
Set a Tuple2 parameter on an NLP model object
}
