% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{nlp_spark_annotation}
\alias{nlp_spark_annotation}
\title{Create a Spark NLP Annotation object inside of Spark}
\usage{
nlp_spark_annotation(
  sc,
  annotatorType,
  begin,
  end,
  result,
  metadata,
  embeddings = NULL
)
}
\arguments{
\item{sc}{A \code{spark_connection}}

\item{annotatorType}{the type of annotation (string)}

\item{begin}{the index of the first character under this annotation (integer)}

\item{end}{the index after the last character under this annotation (integer)}

\item{result}{the main output of the annotation (string)}

\item{metadata}{associated metadata for this annotation (named list)}

\item{embeddings}{vector of embeddings (Array(Float)). Currently unimplemented.}
}
\value{
the Spark NLP Annotation object
}
\description{
This S3 generic is used for a Spark NLP Annotation object that exists inside of
a Spark session.
}
\seealso{
\url{https://nlp.johnsnowlabs.com/docs/en/concepts#annotation}
}
