% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words_cleaner.R
\name{nlp_stop_words_cleaner}
\alias{nlp_stop_words_cleaner}
\title{Spark NLP StopWordsCleaner}
\usage{
nlp_stop_words_cleaner(x, input_cols, output_col, case_sensitive = NULL,
  locale = NULL, stop_words = NULL,
  uid = random_string("stop_words_cleaner_"))
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{case_sensitive}{Whether to do a case sensitive comparison over the stop words.}

\item{locale}{Locale of the input for case insensitive matching. Ignored when caseSensitive is true.}

\item{stop_words}{The words to be filtered out.}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that excludes from a sequence of strings (e.g. the output of a Tokenizer, Normalizer, 
Lemmatizer, and Stemmer) and drops all the stop words from the input sequences.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#stopwordscleaner}
}
