% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t5_transformer.R
\name{nlp_t5_transformer_pretrained}
\alias{nlp_t5_transformer_pretrained}
\title{Load a pretrained Spark NLP T5 Transformer model}
\usage{
nlp_t5_transformer_pretrained(
  sc,
  input_cols,
  output_col,
  task = NULL,
  max_output_length = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{task}{name to give the task being performed}

\item{max_output_length}{the maximum output length}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{T5TransformerModel} model
}
