% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typed-dependency-parser.R
\name{nlp_typed_dependency_parser}
\alias{nlp_typed_dependency_parser}
\title{Spark NLP TypedDependencyParserApproach}
\usage{
nlp_typed_dependency_parser(
  x,
  input_cols,
  output_col,
  n_iterations = NULL,
  conll_u_path = NULL,
  conll_u_read_as = "TEXT",
  conll_u_options = list(format = "text"),
  conll_2009_path = NULL,
  conll_2009_read_as = "TEXT",
  conll_2009_options = list(format = "text"),
  uid = random_string("typed_dependency_parser_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{n_iterations}{number of iterations}

\item{conll_u_path}{path to a file in CoNLL-U format}

\item{conll_u_read_as}{TEXT or SPARK_DATASET}

\item{conll_u_options}{options to pass to the Spark reader}

\item{conll_2009_path}{path to a file in CoNLL 2009 format}

\item{conll_2009_read_as}{TEXT or SPARK_DATASET}

\item{conll_2009_options}{options to pass to the Spark reader}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that is a labeled parser that finds a grammatical relation between two words in a sentence.
Its input is a CoNLL2009 or ConllU dataset.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#typed-dependency-parser}
}
