% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univ_sent_encoder.R
\name{nlp_univ_sent_encoder}
\alias{nlp_univ_sent_encoder}
\title{Spark NLP UniversalSentenceEncoder}
\usage{
nlp_univ_sent_encoder(
  x,
  input_cols,
  output_col,
  dimension = NULL,
  uid = random_string("univ_sent_encoder_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{dimension}{dimension to use for the embeddings}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML transformer that encodes text into high dimensional vectors that can be used for text classification, 
semantic similarity, clustering and other natural language tasks.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#universalsentenceencoder}
}
