% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivekn-sentiment-detector.R
\name{nlp_vivekn_sentiment_detector}
\alias{nlp_vivekn_sentiment_detector}
\title{Spark NLP ViveknSentimentApproach}
\usage{
nlp_vivekn_sentiment_detector(
  x,
  input_cols,
  output_col,
  sentiment_col,
  prune_corpus = NULL,
  feature_limit = NULL,
  unimportant_feature_step = NULL,
  important_feature_ratio = NULL,
  uid = random_string("vivekn_sentiment_detector_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{sentiment_col}{Column with sentiment analysis row’s result for training.}

\item{prune_corpus}{when training on small data you may want to disable this to not cut off infrequent words}

\item{feature_limit}{}

\item{unimportant_feature_step}{}

\item{important_feature_ratio}{}

\item{uid}{A character string used to uniquely identify the ML estimator.}

\item{...}{Optional arguments, see Details.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
  \item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
  a Spark \code{Estimator} object and can be used to compose
  \code{Pipeline} objects.

  \item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
  the NLP estimator appended to the pipeline.

  \item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
  immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that scores a sentence for a sentiment
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#viveknsentimentdetector}
}
