% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vivekn-sentiment-detector.R
\name{nlp_vivekn_sentiment_pretrained}
\alias{nlp_vivekn_sentiment_pretrained}
\title{Load a pretrained Spark NLP model}
\usage{
nlp_vivekn_sentiment_pretrained(
  sc,
  input_cols,
  output_col,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{ViveknSentimentModel} model
}
