% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yake_model.R
\name{nlp_yake_model}
\alias{nlp_yake_model}
\title{Spark NLP YakeModel}
\usage{
nlp_yake_model(
  x,
  input_cols,
  output_col,
  min_ngrams = NULL,
  max_ngrams = NULL,
  n_keywords = NULL,
  stop_words = NULL,
  threshold = NULL,
  window_size = NULL,
  uid = random_string("yake_model_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{min_ngrams}{select the minimum length of an extracted keyword}

\item{max_ngrams}{select the maximum length of an extracted keyword}

\item{n_keywords}{extract the top N keywords}

\item{stop_words}{set the list of stop words}

\item{threshold}{each keyword will be given a keyword score greater than 0. (Lower the score better the keyword) Set an upper bound for the keyword score from this method.}

\item{window_size}{Yake will construct a co-occurence matrix. You can set the window size for the cooccurence matrix construction from this method. ex: windowSize=2 will look at two words to both left and right of a candidate word.}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#yakemodel-keywords-extraction}
}
