% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_nlp_version}
\alias{set_nlp_version}
\title{Set the version of the Spark NLP library to use}
\usage{
set_nlp_version(version)
}
\arguments{
\item{version}{Spark NLP version number to use when starting Spark Session}
}
\description{
Set the version of the Spark NLP library to use
}
