% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{nlp_annotation}
\alias{nlp_annotation}
\title{Spark NLP S3 Annotation object}
\usage{
nlp_annotation(x)
}
\arguments{
\item{x}{a spark_jobj that is an Annotation object or a named list}
}
\value{
a local nlp_annotation object
}
\description{
A Spark NLP annotation S3 object has the following fields:
\itemize{
\item annotatorType: the type of annotation (String)
\item begin: the index of the first character under this annotation (integer)
\item end: the index after the last character under this annotation (integer)
\item metadata: associated metadata for this annotation (Map(String, String))
\item result: the main output of the annotation (String)
\item embeddings: vector of embeddings (Array(Float))
}
}
\details{
See \url{https://nlp.johnsnowlabs.com/docs/en/concepts#annotation}
}
