% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bert-embeddings.R
\name{nlp_bert_embeddings_pretrained}
\alias{nlp_bert_embeddings_pretrained}
\title{Load a pretrained Spark NLP BertEmbeddings model}
\usage{
nlp_bert_embeddings_pretrained(
  sc,
  input_cols,
  output_col,
  case_sensitive = NULL,
  batch_size = NULL,
  dimension = NULL,
  lazy_annotator = NULL,
  max_sentence_length = NULL,
  storage_ref = NULL,
  name = NULL,
  lang = NULL,
  remote_loc = NULL
)
}
\arguments{
\item{sc}{A Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{case_sensitive}{whether to treat the tokens as case insensitive when looking up their embedding}

\item{batch_size}{batch size}

\item{dimension}{the embedding dimension}

\item{lazy_annotator}{use as a lazy annotator or not}

\item{max_sentence_length}{set the maximum sentence length}

\item{storage_ref}{storage reference name}

\item{name}{the name of the model to load. If NULL will use the default value}

\item{lang}{the language of the model to be loaded. If NULL will use the default value}

\item{remote_loc}{the remote location of the model. If NULL will use the default value}
}
\value{
The Spark NLP model with the pretrained model loaded
}
\description{
Create a pretrained Spark NLP \code{BertEmbeddings} model
}
