% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document-assembler.R
\name{nlp_document_assembler}
\alias{nlp_document_assembler}
\title{Document Assembler}
\usage{
nlp_document_assembler(x, input_col = NULL, output_col = NULL,
  id_col = NULL, metadata_col = NULL,
  uid = random_string("document_assembler_"))
}
\arguments{
\item{x}{A Spark connection, pipeline object, or a Spark data frame.}

\item{input_col}{Input column.}

\item{output_col}{Output column.}

\item{id_col}{ID column.}

\item{metadata_col}{Metadata column.}

\item{uid}{UID}
}
\description{
Document Assembler
}
