% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric-delete.R
\name{nlp_symmetric_delete}
\alias{nlp_symmetric_delete}
\title{Spark NLP SymmetricDeleteApproach}
\usage{
nlp_symmetric_delete(
  x,
  input_cols,
  output_col,
  dictionary_path = NULL,
  dictionary_token_pattern = "\\\\S+",
  dictionary_read_as = "LINE_BY_LINE",
  dictionary_options = list(format = "text"),
  max_edit_distance = NULL,
  dups_limit = NULL,
  deletes_threshold = NULL,
  frequency_threshold = NULL,
  longest_word_length = NULL,
  max_frequency = NULL,
  min_frequency = NULL,
  uid = random_string("symmetric_delete_")
)
}
\arguments{
\item{x}{A \code{spark_connection}, \code{ml_pipeline}, or a \code{tbl_spark}.}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{dictionary_path}{path to dictionary of properly written words}

\item{dictionary_token_pattern}{token pattern used in dictionary of properly written words}

\item{dictionary_read_as}{LINE_BY_LINE or SPARK_DATASET}

\item{dictionary_options}{options to pass to the Spark reader}

\item{max_edit_distance}{Maximum edit distance to calculate possible derived words. Defaults to 3.}

\item{dups_limit}{maximum duplicate of characters in a word to consider.}

\item{deletes_threshold}{minimum frequency of corrections a word needs to have to be considered from training.}

\item{frequency_threshold}{minimum frequency of words to be considered from training.}

\item{longest_word_length}{ength of longest word in corpus}

\item{max_frequency}{maximum frequency of a word in the corpus}

\item{min_frequency}{minimum frequency of a word in the corpus}

\item{uid}{A character string used to uniquely identify the ML estimator.}
}
\value{
The object returned depends on the class of \code{x}.

\itemize{
\item \code{spark_connection}: When \code{x} is a \code{spark_connection}, the function returns an instance of a \code{ml_estimator} object. The object contains a pointer to
a Spark \code{Estimator} object and can be used to compose
\code{Pipeline} objects.

\item \code{ml_pipeline}: When \code{x} is a \code{ml_pipeline}, the function returns a \code{ml_pipeline} with
the NLP estimator appended to the pipeline.

\item \code{tbl_spark}: When \code{x} is a \code{tbl_spark}, an estimator is constructed then
immediately fit with the input \code{tbl_spark}, returning an NLP model.
}
}
\description{
Spark ML estimator that is a spell checker inspired on Symmetric Delete algorithm. It retrieves tokens and utilizes
distance metrics to compute possible derived words.
See \url{https://nlp.johnsnowlabs.com/docs/en/annotators#symmetric-spellchecker}
}
