% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-embeddings.R
\name{nlp_word_embeddings_model}
\alias{nlp_word_embeddings_model}
\title{Create a Spark NLP WordEmbeddingsModel}
\usage{
nlp_word_embeddings_model(
  sc,
  input_cols,
  output_col,
  storage_ref = NULL,
  dimension,
  case_sensitive = NULL,
  include_storage = NULL,
  lazy_annotator = NULL,
  read_cache_size = NULL,
  include_embeddings = NULL,
  uid = random_string("word_embeddings_")
)
}
\arguments{
\item{sc}{Spark connection}

\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{storage_ref}{binding to NerDLModel trained by that embeddings}

\item{dimension}{number of word embeddings dimensions}

\item{case_sensitive}{whether to ignore case in tokens for embeddings matching}

\item{include_storage}{include the storage}

\item{lazy_annotator}{boolean for laziness}

\item{read_cache_size}{size for the read cache}

\item{include_embeddings}{whether or not to include word embeddings when saving this annotator to disk (single or within pipeline)}

\item{uid}{unique identifier for this instance}
}
\value{
a Spark transformer WordEmbeddingsModel
}
\description{
This function creates a WordEmbeddingsModel which uses the provided embeddings_ref.
}
