% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word-embeddings.R
\name{nlp_word_embeddings_model}
\alias{nlp_word_embeddings_model}
\title{Create a Spark NLP WordEmbeddingsModel}
\usage{
nlp_word_embeddings_model(
  sc,
  input_cols,
  output_col,
  storage_ref = NULL,
  dimension,
  case_sensitive = NULL,
  include_storage = NULL,
  lazy_annotator = NULL,
  read_cache_size = NULL,
  include_embeddings = NULL,
  uid = random_string("word_embeddings_")
)
}
\arguments{
\item{input_cols}{Input columns. String array.}

\item{output_col}{Output column. String.}

\item{dimension}{number of word embeddings dimensions}

\item{uid}{unique identifier for this instance}

\item{Spark}{connection}

\item{embeddings_ref}{the reference name for the embeddings cache that the model will use}
}
\value{
a Spark transformer WordEmbeddingsModel
}
\description{
This function creates a WordEmbeddingsModel which uses the provided embeddings_ref.
}
